Refactor module student: entity, controller, service, mapper, repository

+++ src/main/java/com/company/university/student/service/StudentService.java
@@ -1,10 +1,14 @@
 package com.company.university.student.service;
 
-import com.company.university.student.domain.StudentStatus;
-import com.company.university.student.dto.*;
+import com.company.university.lecture.application.LectureNotFoundException;
+import com.company.university.lecture.domain.Lecture;
+import com.company.university.lecture.domain.LectureRepository;
 import com.company.university.student.application.StudentMapper;
+import com.company.university.student.application.StudentNotFoundException;
+import com.company.university.student.application.StudentValidator;
 import com.company.university.student.domain.Student;
 import com.company.university.student.domain.StudentRepository;
+import com.company.university.student.dto.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -13,10 +17,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import static com.company.university.student.application.StudentMapper.*;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -24,28 +25,35 @@ import static com.company.university.student.application.StudentMapper.*;
 public class StudentService {
 
     private final StudentRepository studentRepository;
+    private final LectureRepository lectureRepository;
+    private final StudentValidator validator;
 
     public CreateStudentResponse createStudent(CreateStudentRequest request) {
-        Student student = createStudentWithStatusActive(request);
-        Student studentSaved = studentRepository.save(student);
-        return createStudentResponse(studentSaved);
+        if (request == null) {
+            throw new IllegalArgumentException("Request cannot be null");
+        }
+
+        Student student = StudentMapper.toStudent(request);
+        Student savedStudent = studentRepository.save(student);
+        return StudentMapper.createStudentResponse(savedStudent);
     }
 
-    public Set<FindStudentResponse> getStudents() {
+    public List<FindStudentResponse> getStudents() {
         return studentRepository.findAll()
                 .stream()
                 .map(StudentMapper::findStudentResponse)
-                .collect(Collectors.toSet());
+                .toList();
     }
 
     public FindStudentResponse getStudent(Long id) {
         Student student = studentRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Student not found with id: " + id));
+                .orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + id));
 
         return StudentMapper.findStudentResponse(student);
     }
 
     public Page<FindStudentResponse> getStudents(int page, int size, String sortBy, String direction) {
+        validator.validatePaginationAndSorting(page, size, sortBy, direction);
 
         Sort sort = direction.equalsIgnoreCase("desc")
                 ? Sort.by(sortBy).descending()
@@ -57,19 +65,83 @@ public class StudentService {
                 .map(StudentMapper::findStudentResponse);
     }
 
-    @Transactional
-    public UpdateStudentResponse updateStudent(Long id, UpdateStudentRequest request) {
+    public FindStudentWithLecturesResponse findByIdWithLectures(Long id) {
+        Student student = studentRepository.findByIdWithLectures(id)
+                .orElseThrow(() -> new IllegalArgumentException(
+                        "Student with id " + id + " not found"));
 
-        Student student = studentRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Student not found with id: " + id));
-
-        Student saved = studentRepository.save(student);
-
-        return updateStudentResponse(saved, student);
+        return StudentMapper.toFindStudentWithLecturesResponse(student);
     }
 
-    @Transactional
+    public List<FindStudentWithLecturesResponse> findAllWithLectures() {
+        return studentRepository.findAllWithLectures()
+                .stream()
+                .map(StudentMapper::toFindStudentWithLecturesResponse)
+                .toList();
+    }
+
+    public Page<FindStudentWithLecturesResponse> findAllWithLectures(Pageable pageable) {
+        Page<Student> page = studentRepository.findAllWithLectures(pageable);
+
+        return page.map(StudentMapper::toFindStudentWithLecturesResponse);
+    }
+
+    public UpdateStudentResponse updateStudent(Long id, UpdateStudentRequest request) {
+        if (request == null) {
+            throw new IllegalArgumentException("Request cannot be null");
+        }
+
+        if (id == null) {
+            throw new IllegalArgumentException("Student id cannot be null");
+        }
+
+        Student student = studentRepository.findById(id)
+                .orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + id));
+
+        StudentMapper.updateStudentFromRequest(student, request);
+
+        Student saved = studentRepository.save(student);
+        return StudentMapper.updateStudentResponse(saved);
+    }
+
     public void deleteStudent(Long id) {
+        if (!studentRepository.existsById(id)) {
+            throw new StudentNotFoundException("Student not found with id: " + id);
+        }
         studentRepository.deleteById(id);
+    }
+
+    public void addLectureToStudent(Long studentId, Long lectureId) {
+        if (studentId == null || lectureId == null) {
+            throw new IllegalArgumentException("Id student and id lecture are null");
+        }
+
+        Lecture lecture = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new LectureNotFoundException("Lecture not found with id: " + lectureId));
+
+        Student student = studentRepository.findById(studentId)
+                .orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + studentId));
+
+        validator.validateScheduleConflict(studentId, lecture);
+        student.addLecture(lecture);
+        studentRepository.save(student);
+    }
+
+    public void removeLectureFromStudent(Long lectureId, Long studentId) {
+        if (lectureId == null) {
+            throw new LectureNotFoundException("Lecture id is null");
+        }
+        if (studentId == null) {
+            throw new StudentNotFoundException("Student id is null");
+        }
+
+        Lecture lecture = lectureRepository.findById(lectureId)
+                .orElseThrow(() -> new LectureNotFoundException("Lecture not found with id: " + lectureId));
+
+        Student student = studentRepository.findById(studentId)
+                .orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + studentId));
+
+        student.removeLecture(lecture);
+        studentRepository.save(student);
     }
 }
